refactor: move `Coronas` from root to Renderer folder and fix include paths
