+++ source/game_sa/Renderer/Coronas.cpp
