+++ source/game_sa/Renderer/Coronas.h
